Add change to correct typecast

+++ code/scripting/api/libs/tables.cpp
@@ -228,7 +228,7 @@ ADE_INDEXER(l_Tables_SimulatedSpeechOverrides, "number Index", nullptr, "string"
 	if (!ade_get_args(L, "*i", &idx)) {
 		return ade_set_error(L, "s", "");
 	}
-	if (idx > 0 && (unsigned int) idx <= Generic_message_filenames.size()) {
+	if (idx > 0 && (size_t) idx <= Generic_message_filenames.size()) {
 		idx--; //Convert from Lua to C, as lua indices start from 1, not 0
 		return ade_set_args(L, "s", Generic_message_filenames[idx]);
 	}
@@ -242,4 +242,3 @@ ADE_FUNC(__len, l_Tables_SimulatedSpeechOverrides, nullptr, "Number of simulated
 }
 }
 }
-
